+++ server/src/main/java/com/holidayplanner/repository/ResultRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResultRepository extends JpaRepository<Result, Long> {
+public interface ResultRepository extends JpaRepository<Result, Integer> {
     // Find results by journey ID
-    List<Result> findByJourneyId(Long journeyId);
+    List<Result> findByJourneyId(Integer journeyId);
 
     // Find results by vehicle type
     List<Result> findByVehicle(String vehicle);
